feat: split git worktree and repository in separate files
