+++ src/git/mod.rs
@@ -1,0 +1,5 @@
+mod repository;
+mod worktree;
+
+pub use repository::{list_repositories, Repository};
+pub use worktree::{delete_worktree, Worktree};
